requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0 